+++ native-windows-gui/src/controls/frame.rs
@@ -1,8 +1,10 @@
 use winapi::um::winuser::{WS_VISIBLE, WS_DISABLED, WS_BORDER, WS_CHILD, WS_CLIPCHILDREN, WS_EX_CONTROLPARENT};
 use crate::win32::window_helper as wh;
 use crate::win32::base_helper::check_hwnd;
-use crate::NwgError;
+use crate::{NwgError, RawEventHandler, unbind_raw_event_handler, bind_raw_event_handler_inner};
 use super::{ControlBase, ControlHandle};
+use std::cell::RefCell;
+
 
 const NOT_BOUND: &'static str = "Frame is not yet bound to a winapi object";
 const BAD_HANDLE: &'static str = "INTERNAL ERROR: Frame handle is not HWND!";
@@ -37,15 +39,17 @@ Requires the `frame` feature.
   * `enabled`:  If the frame children can be used by the user.
   * `flags`:    A combination of the FrameFlags values.
   * `ex_flags`: A combination of win32 window extended flags. Unlike `flags`, ex_flags must be used straight from winapi
-  * `OnMouseWheel`: Generic mouse wheel event
-
+  
 **Control events:**
   * `MousePress(_)`: Generic mouse press events on the button
   * `OnMouseMove`: Generic mouse mouse event
+  * `OnMouseWheel`: Generic mouse wheel event
+
 */
-#[derive(Default, PartialEq, Eq)]
+#[derive(Default)]
 pub struct Frame {
-    pub handle: ControlHandle
+    pub handle: ControlHandle,
+    handler0: RefCell<Option<RawEventHandler>>,
 }
 
 impl Frame {
@@ -56,6 +60,7 @@ impl Frame {
             position: (0, 0),
             enabled: true,
             flags: None,
+            background_color: None,
             ex_flags: 0,
             parent: None,
         }
@@ -138,17 +143,55 @@ impl Frame {
         WS_CHILD | WS_CLIPCHILDREN
     }
 
+    fn set_background_color(&self, c: [u8;3]) {
+        use winapi::um::winuser::{FillRect, BeginPaint, EndPaint, PAINTSTRUCT};
+        use winapi::um::wingdi::{CreateSolidBrush, RGB};
+        use winapi::um::winuser::WM_PAINT;
+
+        let brush = unsafe { CreateSolidBrush(RGB(c[0], c[1], c[2])) };
+        let handler0 = bind_raw_event_handler_inner(&self.handle, 2000, move |hwnd, msg, _w, _l| {
+            match msg {
+                WM_PAINT => unsafe {
+                    let mut paint: PAINTSTRUCT = ::std::mem::zeroed();
+                    BeginPaint(hwnd, &mut paint);
+                    FillRect(paint.hdc, &paint.rcPaint, brush as _);
+                    EndPaint(hwnd, &paint);
+                    return Some(0);
+                },
+                _ => {}
+            }
+
+            None
+        });
+
+        *self.handler0.borrow_mut() = Some(handler0.unwrap());
+    }
+
 }
 
 impl Drop for Frame {
     fn drop(&mut self) {
+        let handler = self.handler0.borrow();
+        if let Some(h) = handler.as_ref() {
+            drop(unbind_raw_event_handler(h));
+        }
+
         self.handle.destroy();
     }
 }
+
+impl PartialEq for Frame {
+    fn eq(&self, other: &Self) -> bool {
+        self.handle == other.handle
+    }
+}
+
+
 pub struct FrameBuilder {
     size: (i32, i32),
     position: (i32, i32),
     enabled: bool,
+    background_color: Option<[u8; 3]>,
     flags: Option<FrameFlags>,
     ex_flags: u32,
     parent: Option<ControlHandle>
@@ -181,6 +224,11 @@ impl FrameBuilder {
         self
     }
 
+    pub fn background_color(mut self, color: Option<[u8;3]>) -> FrameBuilder {
+        self.background_color = color;
+        self
+    }
+
     pub fn parent<C: Into<ControlHandle>>(mut self, p: C) -> FrameBuilder {
         self.parent = Some(p.into());
         self
@@ -207,6 +255,10 @@ impl FrameBuilder {
             .build()?;
 
         out.set_enabled(self.enabled);
+
+        if let Some(background_color) = self.background_color {
+            out.set_background_color(background_color);
+        }
 
         Ok(())
     }