+++ native-windows-wysiwyg/src/gui.rs
@@ -1,5 +1,7 @@
 mod gui_error;
+mod controls;
 mod widget_box;
+mod project_settings_ui;
 mod object_inspector;
 
 mod gui_builder;